+++ CHANGELOG.md
@@ -3,6 +3,7 @@ UNRELEASED
 - Move canonical repository location to <https://github.com/Roguelazer/ferrous-socks>
 - Change primary branch from `master` to `main`
 - Upgrade to Rust 2024; bump MSRV to 1.88.0
+- Replace `log` with `tracing` and make the syslog stuff more flexible
 
 1.0.8
 -----